fix(comp): adjust styles for `NewRecordForm` and `Calendar` components
The current date-picker in shadcn/ui does not show the latest version.
Because shadcn date-picker is using react-day-picker 8.x, which is not compatible with react 19.
So we are using the one from luca-felix/date-picker for a proper integration.
For further details, please refer to the following links:
- https://daypicker.dev/docs/styling#default-style
- https://date-picker.luca-felix.com/

Styles in `NewRecordForm` was fixed to the correct colors.